refactor: LightGodwokenProvider with new argument `network`
