+++ packages/light-godwoken/src/config/index.ts
@@ -1,0 +1,5 @@
+export * from "./types";
+export * from "./utils";
+export * from "./config";
+export * from "./advanced";
+export * from "./predefined";
