+++ packages/light-godwoken/src/config/predefined.ts
@@ -1,0 +1,8 @@
+import { GodwokenNetwork, LightGodwokenConfigMap } from "./types";
+import testnet from "./predefined/testnet";
+import mainnet from "./predefined/mainnet";
+
+export const predefinedConfigs: Record<GodwokenNetwork, LightGodwokenConfigMap> = {
+  testnet,
+  mainnet,
+};
