+++ packages/light-godwoken/src/config/types.ts
@@ -59,16 +59,7 @@ export type LightGodwokenConfigMap = {
 
 export type GodwokenVersion = keyof LightGodwokenConfigMap;
 
-export type LightGodwokenTokenType = {
-  id: number;
-  symbol: string;
-  name: string;
-  decimals: number;
-  tokenURI: string;
-  address: string;
-  l1LockArgs: string;
-  layer1UAN: string;
-  layer2UAN: string;
-  layer1DisplayName?: string;
-  layer2DisplayName?: string;
-};
+export enum GodwokenNetwork {
+  Testnet = "testnet",
+  Mainnet = "mainnet",
+}
