+++ packages/light-godwoken/src/lightGodwoken.ts
@@ -1,18 +1,11 @@
-import {
-  Cell,
-  Hash,
-  helpers,
-  HexNumber,
-  HexString,
-  Script,
-  toolkit,
-  utils,
-  BI,
-  core,
-  Transaction,
-} from "@ckb-lumos/lumos";
+import EventEmitter from "events";
+import { debug } from "./debug";
+import isEqual from "lodash.isequal";
 import * as secp256k1 from "secp256k1";
-import { getCellDep } from "./constants/configUtils";
+import { toolkit, utils, core } from "@ckb-lumos/lumos";
+import { Cell, Hash, helpers, HexNumber, HexString, Script, BI, Transaction } from "@ckb-lumos/lumos";
+import { CellDep, CellWithStatus, DepType, OutPoint, Output, TransactionWithStatus } from "@ckb-lumos/base";
+import { getCellDep, getAdvancedSettings, GodwokenVersion, LightGodwokenConfig } from "./config";
 import LightGodwokenProvider from "./lightGodwokenProvider";
 import {
   DepositPayload,
@@ -34,8 +27,6 @@ import {
   SUDT_CELL_CAPACITY,
   GetSudtBalance,
 } from "./lightGodwokenType";
-import { debug } from "./debug";
-import { GodwokenVersion, LightGodwokenConfig } from "./constants/configTypes";
 import {
   DepositCanceledError,
   DepositCellNotFoundError,
@@ -48,10 +39,7 @@ import {
   TransactionSignError,
   WithdrawalTimeoutError,
 } from "./constants/error";
-import { CellDep, CellWithStatus, DepType, OutPoint, Output, TransactionWithStatus } from "@ckb-lumos/base";
-import EventEmitter from "events";
-import isEqual from "lodash.isequal";
-import { getAdvancedSettings } from "./constants/configManager";
+import { getTokenList } from "./tokens";
 
 export default abstract class DefaultLightGodwoken implements LightGodwokenBase {
   provider: LightGodwokenProvider;
@@ -74,6 +62,14 @@ export default abstract class DefaultLightGodwoken implements LightGodwokenBase 
   abstract getVersion(): GodwokenVersion;
   abstract withdrawWithEvent(payload: WithdrawalEventEmitterPayload): WithdrawalEventEmitter;
 
+  getNetwork() {
+    return this.provider.getNetwork();
+  }
+
+  getTokenList() {
+    return getTokenList(this.getNetwork(), this.getVersion());
+  }
+
   // in milliseconds
   getBlockProduceTime(): number {
     return this.provider.getConfig().layer2Config.BLOCK_PRODUCE_TIME * 1000;
@@ -84,9 +80,9 @@ export default abstract class DefaultLightGodwoken implements LightGodwokenBase 
   }
 
   getAdvancedSettings() {
-    const cancelTimeOut = getAdvancedSettings(this.getVersion()).MIN_CANCEL_DEPOSIT_TIME;
+    const advancedSettings = getAdvancedSettings(this.getNetwork(), this.getVersion());
     return {
-      cancelTimeOut,
+      cancelTimeOut: advancedSettings.MIN_CANCEL_DEPOSIT_TIME,
     };
   }
 
@@ -230,7 +226,7 @@ export default abstract class DefaultLightGodwoken implements LightGodwokenBase 
         data: "0x",
       });
     }
-    const { layer2Config, layer1Config } = this.provider.getLightGodwokenConfig();
+    const { layer2Config, layer1Config } = this.provider.getConfig();
     const depositLockDep: CellDep = {
       out_point: {
         tx_hash: layer2Config.SCRIPTS.deposit_lock.cell_dep.out_point.tx_hash,
@@ -269,7 +265,7 @@ export default abstract class DefaultLightGodwoken implements LightGodwokenBase 
   }
 
   async getRollupCellDep(): Promise<CellDep> {
-    const { layer2Config } = this.provider.getLightGodwokenConfig();
+    const { layer2Config } = this.provider.getConfig();
     const result = await this.godwokenClient.getLastSubmittedInfo();
     const txHash = result.transaction_hash;
     const tx = await this.getPendingTransaction(txHash);
@@ -401,7 +397,7 @@ export default abstract class DefaultLightGodwoken implements LightGodwokenBase 
     const outputCell = this.generateDepositOutputCell(collectedCells, freeCapacityProviderCells, payload);
     let txSkeleton = helpers.TransactionSkeleton({ cellProvider: this.provider.ckbIndexer });
 
-    const { layer1Config } = this.provider.getLightGodwokenConfig();
+    const { layer1Config } = this.provider.getConfig();
     txSkeleton = txSkeleton
       .update("inputs", (inputs) => {
         return inputs.push(...collectedCells, ...freeCapacityProviderCells);
