+++ packages/light-godwoken/src/tokens/tokens/testnet.ts
@@ -1,11 +1,6 @@
-import { LightGodwokenTokenType } from "./configTypes";
-import { TOKEN_LIST_V0_MAINNET } from "./godwokenTokensV0";
-import { TOKEN_LIST_V1_MAINNET } from "./godwokenTokensV1";
-import { isMainnet } from "../env";
+import { LightGodwokenToken } from "../constants";
 
-export const CKB_SUDT_ID = 1; // This is default sudt id fro ckb on Godwoken
-
-export const TOKEN_LIST_V0: LightGodwokenTokenType[] = [
+export const TOKEN_LIST_TESTNET_V0: LightGodwokenToken[] = [
   {
     id: 120,
     symbol: "TTKN",
@@ -71,7 +66,7 @@ export const TOKEN_LIST_V0: LightGodwokenTokenType[] = [
   },
 ];
 
-export const TOKEN_LIST_V1: LightGodwokenTokenType[] = [
+export const TOKEN_LIST_TESTNET_V1: LightGodwokenToken[] = [
   {
     id: 80,
     symbol: "TTKN",
@@ -136,78 +131,3 @@ export const TOKEN_LIST_V1: LightGodwokenTokenType[] = [
     layer2DisplayName: "ETH (via Forcebridge from ETH)",
   },
 ];
-
-export const getTokenList = () => {
-  if (isMainnet) {
-    return {
-      v0: TOKEN_LIST_V0_MAINNET,
-      v1: TOKEN_LIST_V1_MAINNET,
-    };
-  }
-  return {
-    v0: TOKEN_LIST_V0,
-    v1: TOKEN_LIST_V1,
-  };
-};
-
-type UAN = {
-  asset: { assetSymbol: string; chainSymbol: string };
-  route: Array<{ bridgeSymbol: string; chainSymbol: string }>;
-};
-
-const defaultChains: Record<string, string> = {
-  // gw: "Godwoken",
-  ckb: "CKB",
-  eth: "Ethereum",
-  bsc: "BSC",
-};
-const defaultBridges: Record<string, string> = {
-  // gb: "Godwoken Bridge",
-  fb: "Force Bridge",
-};
-
-const UAN_REGEX = /^(\w+\.\w+)(\|\w+\.\w+)*$/;
-
-export function parse(uan: string): UAN {
-  if (!UAN_REGEX.test(uan)) throw new Error("Invalid UAN: " + uan);
-
-  const [assetPart, ...pathsPart] = uan.split("|");
-  const [assetSymbol, assetChainSymbol] = assetPart.split(".");
-
-  const route = pathsPart.map((path) => {
-    const [bridgeSymbol, chainSymbol] = path.split(".");
-    return { bridgeSymbol, chainSymbol };
-  });
-
-  return {
-    asset: { assetSymbol, chainSymbol: assetChainSymbol },
-    route: route,
-  };
-}
-
-/**
- *
- * @param uan UAN or UAN string
- * @param mapping bridges and chains to translate, by default use {defaultBridges} and {defaultChains}
- * @returns human-readable uan
- */
-export function translate(
-  uan: string | UAN,
-  mapping: { bridges: Record<string, string>; chains: Record<string, string> } = {
-    bridges: defaultBridges,
-    chains: defaultChains,
-  },
-): string {
-  const { asset, route } = typeof uan === "string" ? parse(uan) : uan;
-  const { bridges, chains } = mapping;
-
-  const filteredRoutes = route.filter(({ chainSymbol, bridgeSymbol }) => bridges[bridgeSymbol] && chains[chainSymbol]);
-
-  let routeDisplayName = filteredRoutes
-    .map(({ chainSymbol, bridgeSymbol }) => `${bridges[bridgeSymbol]} from ${chains[chainSymbol]}`)
-    .join(" and ");
-
-  routeDisplayName = routeDisplayName ? `(via ${routeDisplayName})` : "";
-
-  return `${asset.assetSymbol}${routeDisplayName}`;
-}
