--- packages/light-godwoken/src/utils.ts
@@ -1,3 +1,0 @@
-export const isSpecialWallet = () => {
-  return window && ((window.ethereum as any)?.isSafePal || (window.ethereum as any)?.isAlphaWallet);
-};
