+++ packages/light-godwoken/tsconfig.json
@@ -9,5 +9,8 @@
   "include": [
     "src/**/*",
     "src/**/*.json"
+  ],
+  "exclude": [
+    "**/dist"
   ]
 }
